requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-modeltranslation==0.17.3
django-rosetta==0.9.6
django-summernote==0.8.11.6
idna==2.10
oauthlib==3.1.1
Pillow==8.2.0
polib==1.1.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
